test: migrate more tests
